refactor: move generate feedback to shared repo
