+++ lib/features/review_page/data/datasources/sq3r/sq3r_remote_datasource.dart
@@ -15,75 +15,6 @@ class Sq3rRemoteDataSource {
 
   const Sq3rRemoteDataSource(this._firestore, this._auth);
 
-  Future<String> getSq3rFeedback(
-      String noteContextWithSummary, String questionAndAnswers) async {
-    final systemMessage = OpenAIChatCompletionChoiceMessageModel(
-      content: [
-        OpenAIChatCompletionChoiceMessageContentItemModel.text(
-          """
-           As a helpful assistant you will help the student analyze his/her given answers.
-           Give constructive feedback and return a JSON with the properties:
-           
-           "acknowledgement" as a string
-           "missed" as as string
-           "suggestions" as string
-           "isValid" as boolean 
-          """,
-        ),
-      ],
-      role: OpenAIChatMessageRole.system,
-    );
-
-    final userMessage = OpenAIChatCompletionChoiceMessageModel(
-      content: [
-        OpenAIChatCompletionChoiceMessageContentItemModel.text(
-          """
-          Provide constructive feedback for a student based on the following context:
-
-          The student's note with his/her summary/key points made: $noteContextWithSummary
-          Student's Answers to Questions: $questionAndAnswers
-          
-          Based on the above information, generate feedback in JSON that:
-          
-          Acknowledges the student's correct points as "acknowledgement".
-          Identifies any key points or details the student missed or misunderstood as "missed".
-          Offers suggestions for improvement, including specific areas to review or additional context for better understanding as "suggestions".
-          And include a property "isValid" if the content given is valid and not just random letters or gibberish.
-          
-          
-          Structure the feedback to be encouraging and actionable, guiding the student toward refining their understanding of the material.
-          """,
-        ),
-      ],
-      role: OpenAIChatMessageRole.user,
-    );
-
-    final requestMessages = [
-      systemMessage,
-      userMessage,
-    ];
-
-    OpenAIChatCompletionModel chatCompletion =
-        await OpenAI.instance.chat.create(
-      model: "gpt-4o-mini",
-      responseFormat: {"type": "json_object"},
-      messages: requestMessages,
-      temperature: 0.2,
-      maxTokens: 600,
-    );
-
-    String? completionContent =
-        chatCompletion.choices.first.message.content!.first.text;
-
-    var decodedJson = json.decode(completionContent!);
-
-    if (!decodedJson['isValid']) {
-      throw "Given note is not valid, Please try again.";
-    }
-
-    return completionContent;
-  }
-
   Future<String> saveQuizResults(Sq3rModel sq3rModel) async {
     var userId = _auth.currentUser!.uid;
